fix: rendering latency
reduced rendering latency by avoiding page components that are `async` functions (server component)

the most popular section renders two times in a same page (depending of the screen size, but one is hidden when the other is visible). It should be avoided to limit fetch requests
